Refactor search functionality, debounce search request, add loader component
