+++ package.json
@@ -39,6 +39,7 @@
     "isomorphic-fetch": "^2.2.1",
     "jsonwebtoken": "^7.3.0",
     "kue": "^0.11.5",
+    "lodash": "^4.17.4",
     "node-cron": "^1.1.3",
     "pg-promise": "^5.6.4",
     "prop-types": "^15.5.6",
